Add: Buggy CAN code and can data structure

+++ can.h
@@ -10,8 +10,9 @@ This driver assumes CANopen standard message ID format.
 #define CAN_ID_NODE2 0x82
 #define CAN_ID_TIME 0x100
 
-typedef struct{
-	uint16_t id;
+typedef struct {
+	uint8_t id;
 	uint8_t data_length;
 	uint8_t data[8];
 } can_message_t;
+
